pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.45.2",
    "hmmlearn>=0.3.0",
    "numpy>=1.22.0,<2.0.0",
    "openai>=1.61.0",
    "pandas>=2.2.0",
    "plotly>=6.0.0",
    "requests>=2.32.3",
    "scikit-learn>=1.6.1",
    "sentence-transformers==2.2.2",
    "streamlit>=1.41.1",
    "tensorflow-cpu>=2.14.0",
    "tensorflow==2.14.0",
    "torch>=2.6.0",
    "transformers==4.36.2",
    "twilio>=9.4.4",
    "yfinance>=0.2.52",
    "trafilatura>=2.0.0",
    "databutton>=0.38.37",
    "firebase-admin>=6.6.0",
    "pyrebase4>=4.8.0",
    "shap>=0.46.0",
    "xgboost>=2.1.3",
    "tweepy>=4.15.0",
    "textblob>=0.19.0",
    "vadersentiment>=3.3.2",
    "faiss-cpu>=1.10.0",
    "dask[distributed]>=2025.1.0",
    "statsmodels>=0.14.4",
    "scipy>=1.15.1",

    "ccxt>=4.4.52",
    "ratelimit>=2.2.1",
    "keras-rl2>=1.0.5",
    "gym>=0.26.2",
    "google-cloud-bigquery>=3.29.0",
    "google-cloud-storage>=3.0.0",
    "pandas-gbq>=0.26.1",
    "pytrends>=4.9.2",
    "praw>=7.8.1",
    "fastapi>=0.115.8",
    "uvicorn>=0.34.0",
]

[[tool.uv.index]]
explicit = true
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"

[tool.uv.sources]
torch = [{ index = "pytorch-cpu", marker = "platform_system == 'Linux'" }]
torchvision = [{ index = "pytorch-cpu", marker = "platform_system == 'Linux'" }]
